Fix invalid atomic orderings found via loom
